bug fix delete folders and sub folders

+++ app/Http/Handlers/FoldersHandler.php
@@ -10,6 +10,7 @@ namespace App\Http\Handlers;
 
 
 use App\Models\Folder\Folder;
+use Illuminate\Http\Response;
 use Illuminate\Support\Facades\DB;
 
 class FoldersHandler
@@ -105,6 +106,9 @@ class FoldersHandler
             $folder = DB::table(self::FOLDERS_TABLE)
                 ->where('id', $id)
                 ->first();
+            if ( $folder === null) {
+                return response('The folder does not exist', 400);
+            }
         } catch (\Exception $e) {
             return response('FoldersHandler: There is something wrong with the database connection', 403);
         }
@@ -149,42 +153,54 @@ class FoldersHandler
         return $this->getFolderById($id);
     }
 
-    public function deleteFolder(Folder $folder)
+    /**
+     * @param Folder | Response $folder
+     * @return Response|Folder
+     */
+    public function deleteFolder($folder)
     {
-        try {
-            $this->documentsHandler->deleteDocumentsByFolderId($folder->getId());
+        if( $folder instanceof Folder ) {
+            if( $folder->getSubFolders() !== null ) {
+                // Delete the subFolders and the linked documents
+                foreach ($folder->getSubFolders() as $subFolder) {
+                    $this->deleteFolder($subFolder);
+                }
+            }
+            try {
+                $this->documentsHandler->deleteDocumentsByFolderId($folder->getId());
 
-            DB::table(self::FOLDERS_LINK_TABLE)
-                ->where('folderId', '=', $folder->getId())
-                ->delete();
-            DB::table(self::FOLDERS_TABLE)->delete($folder->getId());
-        }catch (\Exception $e) {
-            var_dump('delete exception');
-            return response('FoldersHandler: There is something wrong with the database connection', 403);
+                DB::table(self::FOLDERS_LINK_TABLE)
+                    ->where('folderId', $folder->getId())
+                    ->orWhere('folderSubId', $folder->getId())
+                    ->delete();
+                DB::table(self::FOLDERS_TABLE)->delete($folder->getId());
+            }catch (\Exception $e) {
+                var_dump($e->getMessage(), $folder->getId());
+                return response('FoldersHandler: There is something wrong with the database connection', 403);
+            }
         }
-        return true;
+        return $folder;
     }
 
-    public function deleteFolderByProjectId(Int $projectId)
+    /**
+     * @param Response | Folder[] $folders
+     * @return Response | Folder[]
+     */
+    public function deleteFolders($folders)
     {
-        try {
-            $folders = $this->getFoldersByProjectId($projectId);
-            foreach ($folders as $folder) {
-                if( $folder->getSubFolders() !== null ) {
-                    // Delete the subFolders and the linked documents
-                    foreach ($folder->getSubFolders() as $subFolder) {
-                        $this->deleteFolder($subFolder);
-                    }
-                }
+        if ($folders instanceof Response) {
+            return $folders;
+        }
 
+        try {
+            foreach ($folders as $folder) {
                 $this->deleteFolder($folder);
             }
         } catch (\Exception $e) {
-            var_dump($e->getMessage());
             return response('FoldersHandler: There is something wrong with the database connection', 403);
         }
 
-        return true;
+        return $folders;
     }
 
     /**
